Added "select on lcick" function for Richtextboxes
hi there...
I added the function that text is 'marked' when lcicked on in a
richtextbox.
Furthermore i quickly added the reset functionality...
Also, i switched code to update the settings GUI with current settings
into a routine "settings_reload" which is called after loading the
settings form (and hsould be called everystime when settings are
changed)
The function to mark clicked text is only active in the settings form
currently. laos there is no edit functinality yet and i havent added the
starttime functiion yet - but already preparated the GUI for it.
Hope you like it - feel free to change it hoever you want. but the code
should held anyway.

Great day mate!