Added Check_Application_Autostart Function
Added function to check the autostart registry key and to rewrite it if needed. 
Not tested yet! (Don't have Viusal Studio right now to debug it!)
Also the function is not called yet anywhere in the code! 
I'll implement it fully when I'm home sometimes... (This function should be called on Startup of the Application - while it's running you cannot Change it anyway.)