Polishing for Release
Polishing for initial Release.
Also heavy changes to the update mechanics - it's now updating over
github - sadly it currently needs editing for each release - I'll try to
fix that for the next release.
Also I reverted the button to close the timetable form so that there is
only the "Apply times" button which writes the settings into the array
and closes thee form at the same time.
Of yourse I also edited the Version Numbers to match 0.0.1.0
I also moved the "default.xml" File into a "Settings" folder (Maybe we
want to have multiple setting files at some point?)And I fixxed some
Bugs I found when adding a directory which already get's backupped when
the user aborts. (The array would get corrupted because only the
sourcepath was entered and added to the arary, then it was aborted -
should be fixxed)
Have a look at the Release Now and test the Update...
I think U will like the whole package =).
Nice work Dude!