added png versions of our icons
added png versions of our icons.
I need them for testing with windows 8. I want to see if it creates a
"Tile" in the start screen. If not I know about a cmdline Tool we could
utilize togheter with those PNG's to add a Tile ourselfes. (The App
could check on startup if a Tile exist, (if its win 8+) and if not add
it.)