Added Distribution_Files Directory
Added Distribution Files - The Fiöles I use to Zip when creating a
Release.
The _Install.bat is new. - we can query the reg key it creates - if it
doesnt we can call this batch and delete it aferwards.
(It create sshortcuts on WinXP and higher... tested with Win8 up to
now... settings up Win10 VMs to test it on there...