added "install" routines for Win8, 8.1 and Win10
added "install" routines for Win8, 8.1 and Win10
Win 10 is a bit buggy thought. (because the tool doing the work for me
is buggy in Win10)
like I said we could call this script from within the application if not
installed. It then copies Shortcuts onto the desktop % start menu
folder. (And in Win8+ creates Tiles)
I have some VM's , you can have alook at it at school if u want to.
Also I can move all needed Files in another directory if U want. (in the
end we can delete those files - it s just for the initial isntallation)