Finished "Restoreoverview.xml" example
Daaaaaaamn it was hard.
The xml part took me about, roughly 4 hours.
Writing somehintg SOMEWHERE into it is easy. finding the best way to
manipulate them was cray. I still have to do lot's of cleaning up.
Also testing, but it seems to work after about 5 mins of testing.
Restore view is not updating correctly yet... working on that next....