Fixxed #45 and minor fixxes
This should fix the Bug.
At least IT doesn't crash If I delete the REstore OVerview. (Which
should be enough to test the case about restoreoverview)
Let me know if it works on your side too.
Nice troubleshooting bro !
If it works it was exactly what yuo thought =)
Also fixxed a Bug where it would report errors of files (that theey
aren't copied) even though they were - the catch? Errorcode 0 AND 1 are
"OK" of Robocopy =). I only had checked for 0...

Have a good one - and let me know if it fixxes the Issues.
(BTW Deleting/Renaming the Debug Dir is NOT a valid way of testing. This
will, for example, break the update function - because the called exe is
not there in the new directory... just if you sometimes try the update
after renaming ;)