Restore function finished!
Hey there... was a long time but hey - the Restore works now.
Will have to test it in depth. But for now it seems ot workp retty
smooth.
Hope you have a look.
Also I added a "Debug Mode" Option, to get more Logs. (Or to say it a
bit more precise - to not get spammed in the logs :P)
Have a good one - hope you like it.
(Still have to do code cleanup and integrate the debug mode compleetely
=) )